% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixTools.R
\name{addOnePoint}
\alias{addOnePoint}
\title{Add data point}
\arguments{
\item{mat}{from makeMatrix}

\item{groups}{from makeGroups}

\item{seed}{passed to set.seed}
}
\description{
Adds one point to each group in the matrix.
}
\details{
This function is designed to add one point in each dimension per group to the
current matrix. The points that are added are done so with a call to
\code{\link{makeMatrix}} where the \code{x} and \code{y} arguments are
identical to the origional matrix. The function is primairly designed to work
with the \code{\link{zeroOrHundred}} function.
}
\examples{
mat <- makeMatrix(x=c(0,100), y=c(0, 100), dim=2, points=10)
groups <- makeGroups(mat, names=c("grp1", "gpr2"))
matPlus <- addOnePoint(mat, groups)
}
\author{
Jason Serviss
}
\keyword{addOnePoint}
