% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixTools.R
\name{checkIdenticalDims}
\alias{checkIdenticalDims}
\title{Check Identical Dimensions}
\arguments{
\item{mat}{\code{\link{makeMatrix}} function output.}

\item{groups}{\code{\link{makeGroups}} function output.}

\item{ident}{A boolean value. Do not alter.}
}
\description{
This function takes an input matrix and checks if any
single dimension for a single group is equal to any other
single dimension for a single group including itself.
For example, with 2 groups with 2 dimensions each, it checks
if x dimension for group 1 is exactly equal to dimension y for
group 1 or dimension x or y for group 2.
\strong{Returns TRUE if none are not identical and FALSE if they are}.
}
\examples{
mat <- makeMatrix(x=c(0,1), y=c(0,1), dim=2, points=10)
groups <- makeGroups(mat, names=c("grp1", "grp2"))
checkIdenticalDims(mat, groups)

mat <- list(matrix(rep(1,40), ncol=2))
groups <- makeGroups(mat, c("grp1", "grp2"))
checkIdenticalDims(mat, groups)
}
\author{
Jason Serviss
}
\keyword{checkIdenticalDims}

