% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixTools.R
\name{checkIdenticalPoints}
\alias{checkIdenticalPoints}
\title{Check Identical Points}
\arguments{
\item{mat}{a matrix from makeMatrix function.}

\item{groups}{a character vector from makeGroups function.}

\item{dim}{Which dimension of the matrix to check.}

\item{identicalPointsThreshold}{Sets the threshold of identical points.}
}
\description{
This function takes an input matrix and checks if, for each group, the frequency
of identical data points is greater than what is specified by the \emph{identicalPointsThreshold}
argumet. If the number of identical points are less than the identicalPointsThreshold
the function returns TRUE and, otherwise, returns FALSE.
}
\examples{
mat <- makeMatrix(x=c(0,1), y=c(0,1), dim=2, points=10, seed=11)
groups <- makeGroups(mat, names=c("grp1", "grp2"))
checkIdenticalPoints(mat, groups, 1, 2)

mat <- list(rbind(mat[[1]], mat[[1]][20,], mat[[1]][20,]))
groups <- makeGroups(mat, c("g1", "g2"))
checkIdenticalPoints(mat, groups, 1, 2)
}
\author{
Jason Serviss
}
\keyword{checkIdenticalPoints}

