% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalTools.R
\name{extendToMatch}
\alias{extendToMatch}
\title{Used if LHS is larger than RHS}
\arguments{
\item{l}{argument 1}

\item{r}{argument 2}
}
\description{
Used if LHS is larger than RHS
}
\examples{


cat("no current examples")


}
\author{
Jason Serviss
}
\keyword{extendToMatch}
