% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldFunctions.R
\name{forceUnique}
\alias{forceUnique}
\title{Generate Unique Matrix}
\arguments{
\item{mat}{The matrix to be checked. Output from makeMatrix.}

\item{groups}{Character vector of group names. Can be generated with makeGroups.}

\item{otherMatrices}{an array of the other matrices to be checked against.}

\item{ident}{A boolean value. Do not alter.}

\item{tries}{Equals 1. Do not alter.}
}
\description{
The input matix is re-generated until these constraints of
the identicalChecks function are met.
}
\examples{

mat <- makeMatrix(x=c(0,100), y=c(0,100), dim=2, points=10)
groups <- makeGroups(mat[[1]], names=c("grp1", "grp2"))

otherMatrices <- array(c(rep(0, 40)), dim=c(20, 2, 1))
otherMatrices[,,1] <- makeMatrix(
    x=c(0,1),
    y=c(0,1),
    dim=2,
    points=10
)[[1]]

mat <- forceUnique(
     mat,
     groups,
     otherMatrices
)

}
\author{
Jason Serviss
}
\keyword{forceUnique}

