% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalTools.R
\name{g}
\alias{g}
\title{Grouping the left hand side}
\arguments{
\item{...}{argument 1}
}
\description{
Grouping the left hand side
}
\examples{


cat("no current examples")


}
\author{
Jason Serviss
}
\keyword{g}
