% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hemCancData}
\alias{hemCancData}
\title{tSNE results dataset.}
\format{data.frame
\describe{
\item{X1}{The first tSNE dimension.}
\item{X2}{The second tSNE dimension.}
\item{X3}{The third tSNE dimension.}
\item{geo_accession}{The sample GEO accession number.}
\item{characteristics_ch1.1}{The samples cancer type.}
}}
\usage{
hemCancData
}
\value{
data.frame
}
\description{
tSNE results from the hematological malagnancies anaysis.
}
\examples{
hemCancData
}
