% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotellings.R
\name{hotellingsNA20}
\alias{hotellingsNA20}
\title{Hotellings normalArray test 20 dims}
\description{
Hotellings normalArray test 20 dims
}
\examples{

#hotellingsNA20()


}
\author{
Jason Serviss
}
\keyword{hotellingsNA20}

