% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldFunctions.R
\name{identicalChecks}
\alias{identicalChecks}
\title{Identical Checks}
\arguments{
\item{mat}{The matrix to be checked.}

\item{groups}{Character vector of group names. Can be generated with makeGroups.}

\item{otherMatrices}{an array of the other matrices to be checked against.}

\item{ident}{A boolean value. Do not alter.}
}
\description{
Wrapper for checkIdenticalDims and checkIdenticalMatrix. See the documentation
of the individual functions for more information.
}
\examples{

mat <- makeMatrix(x=c(0,1), y=c(0,1), dim=2, points=10)[[1]]
groups <- makeGroups(mat, names=c("grp1", "grp2"))

otherMatrices <- array(c(rep(0, 40)), dim=c(20, 2, 1))
otherMatrices[,,1] <- makeMatrix(
    x=c(0,1),
    y=c(0,1),
    dim=2,
    points=10
)[[1]]

identicalChecks(mat, groups, otherMatrices)

}
\author{
Jason Serviss
}
\keyword{identicalChecks}

