% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixTools.R
\name{interspersion}
\alias{interspersion}
\title{Calculate interspersion}
\arguments{
\item{mat}{The mattrix to be checked. Usually a \code{\link{makeMatrix}} matrix.}

\item{groups}{A character vector indicating the groups location within the matrix.
Can be generated with the makeGroups function.
Can only accept 2 groups.}

\item{dim}{Which dimention of the matrix to check.}
}
\description{
Calculates the interspersion between 2 groups. For example, if
the the groups are alternating when the points are ordered, the
interspersion will be 100%. Alternativley, if one groups points
lay within the other groups minimum point and its minimum - 1
point, the result will be 0%. See the examples.
}
\examples{
mat <- makeMatrix(x=c(0,1), y=c(0,1), dim=2, points=10, seed=15)
groups <- makeGroups(mat, names=c("g1", "g2"))
interspersion(mat, groups, 1)

mat <- makeMatrix(x=c(0,1), y=c(0.9,0.95), dim=2, points=10, seed=15)
groups <- makeGroups(mat, names=c("g1", "g2"))
interspersion(mat, groups, 1)
}
\author{
Jason Serviss
}
\keyword{interspersion}

