% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lncGenes}
\alias{lncGenes}
\title{lncRNAs dataset.}
\format{character vector}
\usage{
lncGenes
}
\value{
character vector
}
\description{
A character vector of lncRNA ensembl IDs included in the hematological cancer
analysis.
}
\examples{
lncGenes
}
