% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixTools.R
\name{makeGroups}
\alias{makeGroups}
\title{Make Groups}
\arguments{
\item{mat}{The matrix for which the groups should be generated.}

\item{names}{A character vector indicating the names of the groups.}

\item{...}{Arguments to pass on.}
}
\description{
This function makes the \code{groups} input variable to ClusterSignificance from a
\code{\link{makeMatrix}} generated matrix.
}
\details{
This function is specifically designed for the generation of the groups argument corresponding
to the output from the \code{\link{makeMatrix}} function. It always results in 2 groups with
each group being assigned an equal amount of points in the matrix.
}
\examples{
mat <- makeMatrix(x=c(0,1), y=c(0,1), dim=2, points=10)
groups <- makeGroups(mat, names=c("grp1", "grp2"))
}
\author{
Jason Serviss
}
\keyword{makeGroups}

