% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixTools.R
\name{makeMatrix}
\alias{makeMatrix}
\title{Make Matrix}
\arguments{
\item{x}{A numerical vector of length 2 specifying the range of values
possible for group 1.}

\item{y}{A numerical vector of length 2 specifying the range of values
possible for group 2.}

\item{dim}{The number of dimensions (columns) to generate.}

\item{points}{The number of data points in each group.}

\item{by}{Indicates the frequency of ponts which are subsequently randomly
picked according to the \emph{points} argument.}

\item{seed}{Sets the seed to make results reproducible.}

\item{replace}{Passed to the \code{\link[base]{sample}} function.}

\item{...}{Arguments to pass on.}
}
\description{
This function is specifically designed for the generation of 2 groups of
points within a matrix.
}
\details{
Makes a matrix containing data for two groups of points.
Specifying a specific number of points, dimensions, and range
of points for each group is suported. The resulting matrix will have group
1 occupying the top half of the matrix and group 2 the bottom half.
}
\examples{
mat <- makeMatrix(x=c(0,1), y=c(0,1), dim=2, points=10)
}
\author{
Jason Serviss
}
\keyword{makeMatrix}
