% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixTools.R
\name{makeMatrixNorm}
\alias{makeMatrixNorm}
\title{Make Normal Matrix}
\arguments{
\item{mean}{A numerical vector of length 2 specifying the mean values for each group.}

\item{sd}{A numerical vector of length 2 specifying the standard deviation for each group.}

\item{dim}{The number of dimensions (columns) to generate.}

\item{points}{The number of data points in each group.}

\item{seed}{Allows the user to set the seed previous to calling rnorm.}

\item{...}{Arguments to pass on.}
}
\description{
This function is specifically designed for the generation of 2 groups of points within a matrix where each dimension for each group is normally distributed.
}
\details{
Makes a matrix with containing data for two groups of points.
Specifying a specific number of points, dimensions, and mean
and standard deviation for all groups and dimensions is suported.
The resulting matrix will have group 1 occupying the top half of
the matrix and group 2 the bottom half.
}
\examples{

mat <- makeMatrixNorm(mean=c(0,0), sd=c(1,1), dim=2, points=1000)


}
\author{
Jason Serviss
}
\keyword{makeMatrixNorm}

