% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalTools.R
\name{\%=\%.lbunch}
\alias{\%=\%.lbunch}
\title{Binary Operator}
\arguments{
\item{l}{argument 1}

\item{r}{argument 2}

\item{...}{argument 3}
}
\description{
Binary Operator
}
\examples{


cat("no current examples")


}
\author{
Jason Serviss
}
\keyword{internal}

