% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hematologicalCancers.R
\name{normalMDS}
\alias{normalMDS}
\title{normal vs. MDS}
\arguments{
\item{...}{Arguments to pass on.}
}
\description{
Plots the clustering of the normal and MDS cancer types.
}
\examples{
\dontrun{normalMDS()}
}
\author{
Jason Serviss
}
\keyword{normalMDS}
