% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldFunctions.R
\name{overlapSpecificMatrix}
\alias{overlapSpecificMatrix}
\title{Overlap Specific Matrix}
\arguments{
\item{desiredOverlap}{The desired overlap between group 1 and 2.}

\item{dim}{The number of dimensions (columns) to generate. Passed to the makeMatrix function.}

\item{points}{The number of data points in each group. Passed to the makeMatrix function.}

\item{...}{Arguments to pass on.}
}
\description{
This function is specifically designed for the generation of 2 groups of points within the matrix which have
a specific overlap.
}
\details{
A wrapper function for the makeMatrix function which allows definition of the resulting overlap between groups.
The overlap is constrained in all dimensions, i.e. if the generated matrix has 2 dimensions and 100% overlap is
specified, the resulting matrix will have 100% overlap of group 1/2 with group 2/1 in both dimensions. Note
that a matrix where group 1 overlaps group 2 x% in dim 1 and group 2 overlaps group 1 x% in dim 2, is valid.
The x and y arguments to makeMatrix are dynamically set within the function by the table resulting from
overlapFunction.
}
\examples{

cat("no example")


}
\author{
Jason Serviss
}
\keyword{overlapSpecificMatrix}

