% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{overlapTable}
\alias{overlapTable}
\title{Run with default settings.}
\format{data.frame}
\usage{
overlapTable
}
\value{
A table describinf how to set x1 to achieve a specific overlap.
}
\description{
Output from overlapProbability function.
}
\examples{
overlapTable
}
