% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalTools.R
\name{parallelCS}
\alias{parallelCS}
\title{Parallelize ClusterSignificance permutation}
\arguments{
\item{mat}{matrix to ClusterSignificance}

\item{groups}{groups to ClusterSignificance}

\item{iter}{iter to ClusterSignificance}

\item{projMethod}{projMethod to ClusterSignificance}

\item{cores}{the number of cores to run parallel}
}
\description{
Only works, for now, with 2 groups.
}
\examples{


cat("no current examples")


}
\author{
Jason Serviss
}
\keyword{parallelCS}

