% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pcpPermutations}
\alias{pcpPermutations}
\title{pcpMatrix permute result.}
\format{PermutationResults}
\usage{
pcpPermutations
}
\value{
PermutationResults
}
\description{
ClusterSignificance permute output from the pcpMatrix dataset.
}
\examples{
pcpPermutations
}
