% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pe.R
\docType{data}
\name{pe}
\alias{pe}
\title{Hematological malagnancies ClusterSignificance result.}
\format{PermutationResults}
\usage{
pe
}
\value{
PermutationResults
}
\description{
ClusterSignificance permute output from the hematological cancer dataset.
}
\examples{
pe
}

