% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixTools.R
\name{permMatrix}
\alias{permMatrix}
\title{Make permutation matrix}
\arguments{
\item{mat}{from makeMatrix}

\item{groups}{from makeGroups}

\item{iterations}{number of iterations.}
}
\description{
The .permMatrix function from ClusterSignificance.
}
\examples{

mat <- makeMatrix(x=c(0,100), y=c(0, 100), dim=2, points=10)
groups <- makeGroups(mat, names=c("grp1", "gpr2"))
matPlus <- addOnePoint(mat, groups)


}
\author{
Jason Serviss
}
\keyword{permMatrix}
