% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/falseNegativeTest.R
\name{plotFNtest}
\alias{plotFNtest}
\title{Plot false negative test}
\description{
Plots the saved results from the false negative test.
}
\examples{

#plotFNtest()


}
\author{
Jason Serviss
}
\keyword{plotFNtest}

