% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/falsePositiveTest.R
\name{plotFPtest}
\alias{plotFPtest}
\title{Plot false positive test}
\description{
Plots the saved results from the false positive test.
}
\examples{

#plotFPtest()


}
\author{
Jason Serviss
}
\keyword{plotFPtest}

