% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotellings.R
\name{plotHotellingsNA}
\alias{plotHotellingsNA}
\title{Plot Hotellings normalArray test}
\description{
Plot Hotellings normalArray test
}
\examples{

#hotellingsNA()


}
\author{
Jason Serviss
}
\keyword{plotHotellingsNA}

