% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotellings.R
\name{plotHotellingsNA20}
\alias{plotHotellingsNA20}
\title{Plot Hotellings normalArray 20 dim test}
\description{
Plot Hotellings normalArray 20 dim test
}
\examples{

#hotellingsNA()


}
\author{
Jason Serviss
}
\keyword{plotHotellingsNA20}

