% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotellings.R
\name{plotHotellingsSD}
\alias{plotHotellingsSD}
\title{Plot Hotellings SD test}
\description{
Plot Hotellings SD test
}
\examples{

#plotHotellingsSD()


}
\author{
Jason Serviss
}
\keyword{plotHotellingsSD}

