% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectionTests.R
\name{projectionTest1}
\alias{projectionTest1}
\title{Projection test 1}
\description{
This function performs projection test 1. This is designed to examine
the differences in group overlap pre- and post-projection using the Pcp
and Mlp functions.
}
\details{
To do this we utilize the overlapMatrices dataset which was derived with
the overlapMatrices function. For each matrix in the dataset, with
a known groups overlap, we run the Pcp and Mlp functions projecting
all points into one dimension. We subsequently use the calculateOverlaps
function to calculate the group overlap after projection from the
points.onedim variable. The pre- and post-projection groups overlap
data is then output data is then output in the form of a data frame.
}
\examples{

#projectionTest1()


}
\author{
Jason Serviss
}
\keyword{projectionTest1}

