% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivitySpecificityTests.R
\name{sensitivityTest}
\alias{sensitivityTest}
\title{Sensitivity test}
\arguments{
\item{interval}{The interval of points amounts that should be tested. i.e. 1
   tests points
amounts \code{1:100}, 10 tests \code{seq(1, 100, 10)}, etc.}

\item{iterations}{The number of iterations for each call to
\code{\link[ClusterSignificance]{permute}}.}

\item{cores}{The number of cores the test should be parallalized with.}

\item{save}{Should the output be saved?}

\item{verbose}{Should the function be verbose?}

\item{outPath}{The location the file should be saved to if \code{save=TRUE}.}
}
\description{
This function performs the sensitivity test. This test is designed to
determine the true positive rate of the Pcp and Mlp methods. To do this
we utilize the zero dataset which was derived using the
\code{\link{zeroOrHundred}} function with argument \code{overlap}=0. The
dataset includes matrices where the groups overlap is 0% and thus we
anticipate that significant seperations should be found with Pcp or Mlp for
all matrices. For each matrice, the permute function is run, using both the
Pcp and Mlp methods afterwhich, the resulting pvalue is recorded.
}
\examples{
\dontrun{sensitivityTest(interval=50, iterations=1, cores=1, save=FALSE)}
}
\author{
Jason Serviss
}
\keyword{sensitivityTest}
