% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sensitivityTestResults}
\alias{sensitivityTestResults}
\title{Sensitivity test results dataset.}
\format{data.frame
\describe{
\item{points}{The number of points per group.}
\item{rep}{The repitition.}
\item{pValuePcp}{The pValue from ClusterSignificance with 10^4 iterations.}
}}
\usage{
sensitivityTestResults
}
\value{
data.frame
}
\description{
Output from the sensitivityTest function with default arguments.
}
\examples{
sensitivityTestResults
}

