% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{sensSpecPlot}
\alias{sensSpecPlot}
\title{Sensitivity and Specificity Figure}
\description{
This function plots the specificity and sensivity testing results.
}
\examples{

sensSpecPlot()


}
\author{
Jason Serviss
}
\keyword{sensSpecPlot}

