% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specificityTestResults.R
\docType{data}
\name{specificityTestResults}
\alias{specificityTestResults}
\title{Specificity test results dataset.}
\format{data.frame
\describe{
\item{points}{The number of points per group.}
\item{rep}{The repitition.}
\item{pValuePcp}{The pValue from ClusterSignificance with 10^4 iterations.}
}}
\usage{
specificityTestResults
}
\value{
data.frame
}
\description{
Output from the specificityTest function with default arguments.
}
\examples{
specificityTestResults
}

