% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hematologicalCancers.R
\name{tsnePlots}
\alias{tsnePlots}
\title{tSNE plots}
\arguments{
\item{...}{Arguments to pass on.}
}
\description{
Performs the tSNE plots for the hematological malagnancies dataset.
}
\examples{
\dontrun{tsnePlots()}
}
\author{
Jason Serviss
}
\keyword{tsnePlots}
