% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hematologicalCancers.R
\name{hematologicalCancers}
\alias{hematologicalCancers}
\title{Hematological Cancers}
\arguments{
\item{iterations}{Number of permutations.}

\item{...}{Arguments to pass on.}
}
\description{
Downloads, processes, clusters, and runs ClusterSignificance on
the hematological malagnancies dataset.
}
\examples{
\dontrun{hematologicalCancers()}
}
\author{
Jason Serviss
}
\keyword{hematologicalCancers}
